# internal/api/docs/swagger.yaml
basePath: /api/v1
...
host: localhost:8080
info:
  contact: {}
  description: Aether Notebooks API — collaborative SQL/data notebook platform
  title: Aether API
  version: 1.0.0
paths:
...
  /api/v1/public/{token}:
    get:
      description: Returns a publicly shared notebook or dashboard by token
      parameters:
      - description: Public sharing token
        in: path
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get public resource
      tags:
      - public
...
  /api/v1/ws/notebooks/{id}:
    get:
      description: WebSocket connection for real-time notebook collaboration
      parameters:
      - description: Notebook ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "101":
          description: Switching Protocols
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Notebook WebSocket
      tags:
      - websocket
...
  /auth/oidc/{provider}/callback:
    get:
      description: Handle the OIDC callback after provider authentication
      parameters:
      - description: OIDC Provider ID
        in: path
        name: provider
        required: true
        type: string
      - description: Authorization code
        in: query
        name: code
        required: true
        type: string
      - description: State parameter
        in: query
        name: state
        required: true
        type: string
      responses:
        "302":
          description: Found
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: OIDC callback
      tags:
      - auth
  /auth/oidc/{provider}/login:
    get:
      description: Initiate an OIDC login flow for the given provider
      parameters:
      - description: OIDC Provider ID
        in: path
        name: provider
        required: true
        type: string
      responses:
        "302":
          description: Found
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: OIDC login
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    description: Enter "Bearer {token}"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"